WIP - rate limit implementation
Hello @rbranson @tsholmes @ndhoule @thehydroimpulse
Here is a first try on the rate implementation for drip.

Right now it is just on the identify call.
Would be glad to have some feedback to see if I’m going on the right
direction.

Also I’m not sure about the re enqueuing mechanics. Any hint would be
welcome.